// src/app/Models/measurement.ts
// export interface Measurements {
//     Arms: number

//     Chest: number

//     Height: number

//     ShoulderWidth: number

//     Thigh: number

//     Waist: number
//     Weight: number
//     DateMeasured:Date
// }

export interface Measurements {
    arms: number;
    chest: number;
    height: number;
    shoulderWidth: number;
    thigh: number;
    waist: number;
    weight: number;
    dateMeasured:Date;
  }